Changes to script and folder name.
I added a bg and bgm to the script, the bg won't appear tho, I think because it's too big so will rescale it, the bgm appears with the Teacher title so might have to change that too. I also renamed the gui folder to match the gui folder of renpy.