Added music and organized folder
I added some music to the main menu, it's an off vocal from PoPiPas 10th single. I've been trying to add a sound to the start button but I'm having a difficult time with it, I've included the sound as well in case any of y'all would know how to.