The gift script {:
okay so this aint really done yet and was made separate from the game. i will eventually integrate in with the game when i redo all the variables, yeah? also i had some problems so if someone could go into the code and check over the comments i left.....that would be appreciated. thank you!